requirements.txt
PyQt5>=5.15
numpy>=1.21
scipy>=1.7
shapely>=1.8
Pillow>=8.0
PyOpenGL>=3.1
requests>=2.28
